requirements.txt
numpy
deepchem
matplotlib==3.3.4
pandas==1.1.5
rdkit_pypi
scikit_learn
torch
tensorflow
